// packages/client/src/generated/agent-v1/types.ts
export interface components {
  schemas: {
// ...
    UserState: {
      connection?: components["schemas"]["UserStateConnection"];
      evm?: components["schemas"]["UserStateEvm"];
      svm?: components["schemas"]["UserStateSvm"];
      preferences?: {
        [key: string]: unknown;
      };
      ext?: {
        [key: string]: unknown;
      };
    };
    UserStateConnection: {
      is_connected?: boolean;
      provider?: string | null;
      provider_label?: string | null;
      auth_method?: string | null;
    };
    UserStateEvm: {
      address?: string | null;
      chain_id?: number | null;
      ens_name?: string | null;
    };
    UserStateSvm: {
      address?: string | null;
      cluster?: string | null;
      wallet_name?: string | null;
      transport?: string | null;
      capabilities?: string[];
    };
// ...
  };
// ...
}